override linguist statistics
Github uses linguist for language statistics discovery. Unfortunately it does not support q language - thus it is ignoring q files. 
As there are a few .sh files in the project it was displaying statistics that enterprise-components are "100% Shell" which is obviously not true.
Definition of linguist-language and linguist-vendored in .gitattributes allows override of the statistics
https://github.com/github/linguist#overrides